make th auth route a

+++ app/Http/Controllers/IdeaController.php
@@ -19,18 +19,24 @@ class IdeaController extends Controller
         $validated = request()->validate([
             'content' => 'required|min:5|max:50'
         ]);
-        Idea::create(
+        $validated['user_id'] = auth()->id();
 
-            [
-                'content' => request()->get('content', ''),
-            ]
-        );
+        Idea::create($validated);
+        // Idea::create(
+
+        //     [
+        //         'content' => request()->get('content', ''),
+        //     ]
+        // );
         return redirect()->route('dashboard')->with('success', ' Idea created Successfully!');
     }
 
 
     public function destroy(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404, "");
+        }
         $idea->delete();
 
         return redirect()->route('dashboard')->with('success', ' Idea DELETED Successfully!');
@@ -39,6 +45,9 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404, "");
+        }
         $editing = true;
         return view('ideas.show', compact('idea', 'editing'));
     }
@@ -46,6 +55,9 @@ class IdeaController extends Controller
 
     public function update(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404, "");
+        }
         $validated = request()->validate([
             'content' => 'required|min:5|max:50'
         ]);
